secp256k1: Decouple precomputation generation.
The current code that generates the various pre-computed values is
conditionally exported via a build tag that is only set when the code
generation is invoked.  That approach was originally necessary because
it relied on access to unexported code to perform the calculations,
however, now that the package has evolved, the exported API now provides
all of the necessary tools.

Consequently, this decouples the code related to generating pre-computed
values as follows:

- Move the code into the main package that is invoked by go generate
- Updates the generation code to decouple it from the package internals
- Removes the no longer necessary genstatics.go file
- Removes the gensecp256k1 build tag that is no longer necessary
